Integrate with implementation without `if` statement
