+++ src/main/java/com/woowacourse/javaracingcar/ConsoleUserInterface.java
@@ -18,27 +18,29 @@ public class ConsoleUserInterface implements UserInterface {
 
     @Override
     public List<String> promptUserNames() {
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
-        String input = scanner.nextLine();
-        String[] splittedNames = RacingcarUtil.splitIntoNames(input);
-        if (!RacingcarUtil.isValidNameInput(splittedNames)) {
-            return onInvalidUserNames();
+        final String promptText = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
+        System.out.println(promptText);
+        List<String> splittedNames = RacingcarUtil.splitIntoNames(scanner.nextLine());
+
+        while (!RacingcarUtil.isValidNameInput(splittedNames)) {
+            System.out.println("잘못된 입력입니다");
+            System.out.println(promptText);
+            splittedNames = RacingcarUtil.splitIntoNames(scanner.nextLine());
         }
 
-        List<String> convertedList = new ArrayList<>();
-        for (String str : splittedNames) {
-            convertedList.add(str);
-        }
-        return convertedList;
+        return splittedNames;
     }
 
     @Override
     public int promptTries() {
-        System.out.println("시도할 회수는 몇회인가요?");
+        final String promptText = "시도할 회수는 몇회인가요?";
+        System.out.println(promptText);
         int input = scanner.nextInt();
 
-        if (!isValidTries(input)) {
-            onInvalidTries();
+        while (!isValidTries(input)) {
+            System.out.println("잘못된 입력입니다");
+            System.out.println(promptText);
+            input = scanner.nextInt();
         }
 
         return input;
@@ -62,16 +64,6 @@ public class ConsoleUserInterface implements UserInterface {
     public void printWinners(List<CarDto> winners) {
         String str = RacingcarUtil.joinCarNames(winners);
         System.out.println(str + "가 최종 우승했습니다.");
-    }
-
-    private List<String> onInvalidUserNames() {
-        System.out.println("잘못된 입력입니다");
-        return promptUserNames();
-    }
-
-    private int onInvalidTries() {
-        System.out.println("잘못된 입력입니다");
-        return promptTries();
     }
 
     private boolean isValidTries(int tries) {
